per-axis radii, compact manifolds, and a hundred dimensions
Roadmap step 3. The engineering works and the ceiling is raised; the headline
claim does not survive measurement, and the reason is the useful part.

Reach replaces the ball-shaped insertion domain. The D* argument never needed
a ball -- for any S contained in B, f(v) >= dist(v, B) -- so the walk now
takes a Reach, and insert_within is the Ball case. A box of half-extents rho
sits inside a ball of radius |rho|, its circumradius, so describing it as a
ball inflated the visited region by the shape's anisotropy.

Manifold makes a compact manifold a type rather than a hand-built Aabb:
sobolev/exponential/finite_rank constructors, domain, walls, aspect,
log_volume, and effective_dimension -- the participation ratio, which is 2.38
at D = 24 and 2.47 at D = 100 for the same Sobolev decay. Manifold::field
wires domain and walls together so they cannot silently disagree.

Budgets are stored in levels, not full subdivisions. A binary layout spends D
levels per subdivision, so the mildest setting of with_prune_subdiv was a
hundred levels at D = 100 and an undecided proof walked a binary tree a
hundred deep. The subdivision setters remain as exact wrappers, so nothing
below D ~ 12 moves. Same for the arena: max_depth x D <= 255 allowed two
subdivisions at D = 100, now u16 plus with_levels; and child_rect took a 1.6
KiB Aabb by value per call, where halve_into writes one scalar.

The negative result: grow_box and box_is_free are sound and work, but a box
cannot be *certified* much past its inscribed size. The Lipschitz test
compares the cell's half-diagonal against the field's clearance; the
half-diagonal is set by the longest axis and the clearance by the shortest, so
an aspect ratio of 1e4 is a factor of 1e4 the proof must close by refinement
alone. It does, halving the shortfall per four levels while the cost doubles
per level. Measured in log-volume, boxes beat balls only at the mildest
anisotropy tested (D = 24, s = 1) and lose by ln 242 at D = 100, s = 2.

Per-axis radii therefore need a certificate that is anisotropic too, which
section 2.3 promised would not be needed. tests/high_d.rs carries the
evidence, including the budget sweep, and does not assert the sign of the gap.

+++ adaptive-distance-field/tests/layout.rs
@@ -32,7 +32,7 @@ struct Report {
   leaves: usize,
   splits: usize,
   slots: usize,
-  levels: u8,
+  levels: u16,
   bytes: usize,
   checksum: f64,
 }
@@ -97,7 +97,7 @@ macro_rules! measure {
       query = query.min(t1.elapsed());
     }
 
-    let (mut slots, mut levels) = (0usize, 0u8);
+    let (mut slots, mut levels) = (0usize, 0u16);
     field.tree.traverse(&mut |n| { slots += n.data.len(); levels = levels.max(n.depth); Ok(()) }).ok();
     let children = <$L as Layout<D>>::CHILDREN;
